datafactory/requirements.txt
azure-mgmt-datafactory == 1.0.0
azure-identity